refactor: вынесены типы и стили звёзд отзыва в отдельные файлы
