+++ src/components/mock/mock-offers.ts
@@ -1,28 +1,5 @@
+import { OfferDto } from '../../types/types';
 import apartment01 from './../../../markup/img/apartment-01.jpg';
-
-export interface OfferDto {
-  id?: string;
-  title: string;
-  type: string;
-  price: number;
-  city: {
-  name: string;
-  // location: {
-  // latitude: number;
-  // longitude: number;
-  // zoom: number;
-  // };
-  // };
-  // location: {
-  // latitude: number;
-  // longitude: number;
-  // zoom: number;
-  };
-  isFavorite: boolean;
-  isPremium: boolean;
-  rating: number;
-  previewImage: string;
-}
 
 export const mockOffers: OfferDto[] = [
   {
@@ -118,3 +95,4 @@ export const mockOffers: OfferDto[] = [
     rating: 3.4,
   },
 ];
+
