+++ src/pages/offer-screen/components/reviews-form.module.css
@@ -1,0 +1,8 @@
+.form__star-image {
+  fill: #ccc;
+}
+
+.form__star-image.filled {
+  fill: #ff9000;
+}
+
