+++ src/types/types.ts
@@ -1,0 +1,68 @@
+export interface CommentDto {
+  id: string;
+  date: string;
+  user: {
+    name: string;
+    avatarUrl: string;
+    isPro: boolean;
+  };
+  comment: string;
+  rating: number;
+}
+
+export interface OfferDetailsDto {
+  id: string;
+  title: string;
+  type: string;
+  price: number;
+  city: {
+    name: string;
+    // location: {
+    //   latitude: number;
+    //   longitude: number;
+    //   zoom: number;
+    // };
+  };
+  // location: {
+  //   latitude: number;
+  //   longitude: number;
+  //   zoom: number;
+  // };
+  isFavorite: boolean;
+  isPremium: boolean;
+  rating: number;
+  description: string;
+  bedrooms: number;
+  goods: string[];
+  host: {
+    name: string;
+    avatarUrl: string;
+    isPro: boolean;
+  };
+  images: string[];
+  maxAdults: number;
+}
+
+export interface OfferDto {
+  id?: string;
+  title: string;
+  type: string;
+  price: number;
+  city: {
+  name: string;
+  // location: {
+  // latitude: number;
+  // longitude: number;
+  // zoom: number;
+  // };
+  // };
+  // location: {
+  // latitude: number;
+  // longitude: number;
+  // zoom: number;
+  };
+  isFavorite: boolean;
+  isPremium: boolean;
+  rating: number;
+  previewImage: string;
+}
